pyproject.toml
[project]
name = "vnpt-ai-rag"
version = "0.1.0"
description = "High-performance RAG pipeline for Vietnamese multiple-choice questions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-huggingface>=0.1.0",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-experimental>=0.3.0",
    "qdrant-client>=1.12.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=5.1.2",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "colorama>=0.4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
